# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedStock AI - Hospital Supply Chain Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  # SKUs
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all active alerts
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [low_stock, expiry, overstock, forecast]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, warning, info]
        - name: resolved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Mark an alert as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"

  # Crisis Coordination
  /crisis/surplus-listings:
    get:
      operationId: listSurplusListings
      tags: [crisis]
      summary: List surplus inventory available in the network
      responses:
        "200":
          description: Surplus listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurplusListing"
    post:
      operationId: createSurplusListing
      tags: [crisis]
      summary: Publish a surplus listing to the network
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurplusListingBody"
      responses:
        "201":
          description: Surplus listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurplusListing"
  /crisis/requests:
    get:
      operationId: listCrisisRequests
      tags: [crisis]
      summary: List all critical item requests in the network
      responses:
        "200":
          description: Crisis requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrisisRequest"
    post:
      operationId: createCrisisRequest
      tags: [crisis]
      summary: Submit a critical item request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCrisisRequestBody"
      responses:
        "201":
          description: Crisis request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrisisRequest"
  /crisis/requests/{id}/update-status:
    post:
      operationId: updateCrisisRequestStatus
      tags: [crisis]
      summary: Update the status of a crisis request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCrisisStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrisisRequest"
...
  /crisis/mode:
    get:
      operationId: getCrisisMode
      tags: [crisis]
      summary: Get current crisis mode status
      responses:
        "200":
          description: Crisis mode status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrisisModeStatus"
    post:
      operationId: toggleCrisisMode
      tags: [crisis]
      summary: Toggle crisis mode on/off
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleCrisisModeBody"
      responses:
        "200":
          description: Crisis mode toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrisisModeStatus"

  # Hospitals
...
components:
  schemas:
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hospitalId:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - hospitalId
        - createdAt
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        hospitalId:
          type: integer
        description:
          type: string
      required:
        - name
        - hospitalId
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [low_stock, expiry, overstock, forecast]
        severity:
          type: string
          enum: [critical, warning, info]
        title:
          type: string
        message:
          type: string
        skuId:
          type: integer
          nullable: true
        skuName:
          type: string
          nullable: true
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        isResolved:
          type: boolean
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - type
        - severity
        - title
        - message
        - isResolved
        - createdAt
...
    SurplusListing:
      type: object
      properties:
        id:
          type: integer
        hospitalId:
          type: integer
        hospitalName:
          type: string
        hospitalLocation:
          type: string
        skuName:
          type: string
        category:
          type: string
        quantity:
          type: number
        unit:
          type: string
        expiryDate:
          type: string
          format: date
          nullable: true
        pricePerUnit:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        validUntil:
          type: string
          format: date
        status:
          type: string
          enum: [available, matched, fulfilled, expired]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - hospitalId
        - hospitalName
        - hospitalLocation
        - skuName
        - category
        - quantity
        - unit
        - validUntil
        - status
        - createdAt
    CreateSurplusListingBody:
      type: object
      properties:
        skuName:
          type: string
        category:
          type: string
        quantity:
          type: number
        unit:
          type: string
        expiryDate:
          type: string
          format: date
        pricePerUnit:
          type: number
        notes:
          type: string
        validUntil:
          type: string
          format: date
      required:
        - skuName
        - category
        - quantity
        - unit
        - validUntil
    CrisisRequest:
      type: object
      properties:
        id:
          type: integer
        hospitalId:
          type: integer
        hospitalName:
          type: string
        hospitalLocation:
          type: string
        skuName:
          type: string
        category:
          type: string
        quantityNeeded:
          type: number
        unit:
          type: string
        urgency:
          type: string
          enum: [critical, high, medium]
        status:
          type: string
          enum: [open, matched, confirmed, transfer_acknowledged, closed]
        notes:
          type: string
          nullable: true
        isPriority:
          type: boolean
        matchedSurplusId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - hospitalId
        - hospitalName
        - hospitalLocation
        - skuName
        - category
        - quantityNeeded
        - unit
        - urgency
        - status
        - isPriority
        - createdAt
        - updatedAt
    CreateCrisisRequestBody:
      type: object
      properties:
        skuName:
          type: string
        category:
          type: string
        quantityNeeded:
          type: number
        unit:
          type: string
        urgency:
          type: string
          enum: [critical, high, medium]
        notes:
          type: string
        isPriority:
          type: boolean
      required:
        - skuName
        - category
        - quantityNeeded
        - unit
        - urgency
    UpdateCrisisStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, matched, confirmed, transfer_acknowledged, closed]
        notes:
          type: string
      required:
        - status
...
    CrisisModeStatus:
      type: object
      properties:
        active:
          type: boolean
        activatedAt:
          type: string
          format: date-time
          nullable: true
        activatedBy:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
      required:
        - active
    ToggleCrisisModeBody:
      type: object
      properties:
        active:
          type: boolean
        reason:
          type: string
      required:
        - active